backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
PyYAML>=6.0
requests>=2.31
pypdf>=4.0
python-docx>=1.1.0
markdown>=3.6
pandas>=2.0
numpy>=1.26
rich>=13.7

httpx>=0.27
